+++ indexing-service/src/main/java/org/apache/druid/indexing/common/config/TaskConfig.java
@@ -24,7 +24,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
 import org.apache.commons.io.IOUtils;
-import org.apache.commons.lang3.EnumUtils;
 import org.apache.druid.common.config.Configs;
 import org.apache.druid.common.utils.IdUtils;
 import org.apache.druid.java.util.common.ISE;
@@ -65,16 +64,6 @@ public class TaskConfig
     }
   }
 
-  // This enum controls processing mode of batch ingestion "segment creation" phase (i.e. appenderator logic)
-  public enum BatchProcessingMode
-  {
-    OPEN_SEGMENTS, /* mmap segments, legacy code */
-    CLOSED_SEGMENTS, /* Do not mmap segments but keep most other legacy code */
-    CLOSED_SEGMENTS_SINKS /* Most aggressive memory optimization, do not mmap segments and eliminate sinks, etc. */
-  }
-
-  public static final BatchProcessingMode BATCH_PROCESSING_MODE_DEFAULT = BatchProcessingMode.CLOSED_SEGMENTS;
-
   private static final Period DEFAULT_DIRECTORY_LOCK_TIMEOUT = new Period("PT10M");
   private static final Period DEFAULT_GRACEFUL_SHUTDOWN_TIMEOUT = new Period("PT5M");
   private static final boolean DEFAULT_STORE_EMPTY_COLUMNS = true;
@@ -111,12 +100,6 @@ public class TaskConfig
   private final boolean ignoreTimestampSpecForDruidInputSource;
 
   @JsonProperty
-  private final boolean batchMemoryMappedIndex;
-
-  @JsonProperty
-  private final BatchProcessingMode batchProcessingMode;
-
-  @JsonProperty
   private final boolean storeEmptyColumns;
 
   @JsonProperty
@@ -137,9 +120,6 @@ public class TaskConfig
       @JsonProperty("directoryLockTimeout") Period directoryLockTimeout,
       @JsonProperty("shuffleDataLocations") List<StorageLocationConfig> shuffleDataLocations,
       @JsonProperty("ignoreTimestampSpecForDruidInputSource") boolean ignoreTimestampSpecForDruidInputSource,
-      @JsonProperty("batchMemoryMappedIndex") boolean batchMemoryMappedIndex,
-      // deprecated, only set to true to fall back to older behavior
-      @JsonProperty("batchProcessingMode") String batchProcessingMode,
       @JsonProperty("storeEmptyColumns") @Nullable Boolean storeEmptyColumns,
       @JsonProperty("encapsulatedTask") boolean enableTaskLevelLogPush,
       @JsonProperty("tmpStorageBytesPerTask") @Nullable Long tmpStorageBytesPerTask
@@ -171,25 +151,7 @@ public class TaskConfig
     );
 
     this.ignoreTimestampSpecForDruidInputSource = ignoreTimestampSpecForDruidInputSource;
-    this.batchMemoryMappedIndex = batchMemoryMappedIndex;
     this.encapsulatedTask = enableTaskLevelLogPush;
-
-    // Conflict resolution. Assume that if batchMemoryMappedIndex is set (since false is the default) that
-    // the user changed it intentionally to use legacy, in this case oveeride batchProcessingMode and also
-    // set it to legacy else just use batchProcessingMode and don't pay attention to batchMemoryMappedIndexMode:
-    if (batchMemoryMappedIndex) {
-      this.batchProcessingMode = BatchProcessingMode.OPEN_SEGMENTS;
-    } else if (EnumUtils.isValidEnum(BatchProcessingMode.class, batchProcessingMode)) {
-      this.batchProcessingMode = BatchProcessingMode.valueOf(batchProcessingMode);
-    } else {
-      // batchProcessingMode input string is invalid, log & use the default.
-      this.batchProcessingMode = BatchProcessingMode.CLOSED_SEGMENTS; // Default
-      log.warn(
-          "Batch processing mode argument value is null or not valid:[%s], defaulting to[%s] ",
-          batchProcessingMode, this.batchProcessingMode
-      );
-    }
-    log.debug("Batch processing mode:[%s]", this.batchProcessingMode);
 
     this.storeEmptyColumns = Configs.valueOrDefault(storeEmptyColumns, DEFAULT_STORE_EMPTY_COLUMNS);
     this.tmpStorageBytesPerTask = Configs.valueOrDefault(tmpStorageBytesPerTask, DEFAULT_TMP_STORAGE_BYTES_PER_TASK);
@@ -206,8 +168,6 @@ public class TaskConfig
       Period directoryLockTimeout,
       List<StorageLocationConfig> shuffleDataLocations,
       boolean ignoreTimestampSpecForDruidInputSource,
-      boolean batchMemoryMappedIndex,
-      BatchProcessingMode batchProcessingMode,
       boolean storeEmptyColumns,
       boolean encapsulatedTask,
       long tmpStorageBytesPerTask
@@ -223,8 +183,6 @@ public class TaskConfig
     this.directoryLockTimeout = directoryLockTimeout;
     this.shuffleDataLocations = shuffleDataLocations;
     this.ignoreTimestampSpecForDruidInputSource = ignoreTimestampSpecForDruidInputSource;
-    this.batchMemoryMappedIndex = batchMemoryMappedIndex;
-    this.batchProcessingMode = batchProcessingMode;
     this.storeEmptyColumns = storeEmptyColumns;
     this.encapsulatedTask = encapsulatedTask;
     this.tmpStorageBytesPerTask = tmpStorageBytesPerTask;
@@ -311,22 +269,6 @@ public class TaskConfig
   }
 
   @JsonProperty
-  public BatchProcessingMode getBatchProcessingMode()
-  {
-    return batchProcessingMode;
-  }
-
-  /**
-   * Do not use in code! use {@link TaskConfig#getBatchProcessingMode() instead}
-   */
-  @Deprecated
-  @JsonProperty
-  public boolean getbatchMemoryMappedIndex()
-  {
-    return batchMemoryMappedIndex;
-  }
-
-  @JsonProperty
   public boolean isStoreEmptyColumns()
   {
     return storeEmptyColumns;
@@ -366,8 +308,6 @@ public class TaskConfig
         directoryLockTimeout,
         shuffleDataLocations,
         ignoreTimestampSpecForDruidInputSource,
-        batchMemoryMappedIndex,
-        batchProcessingMode,
         storeEmptyColumns,
         encapsulatedTask,
         tmpStorageBytesPerTask
@@ -387,8 +327,6 @@ public class TaskConfig
         directoryLockTimeout,
         shuffleDataLocations,
         ignoreTimestampSpecForDruidInputSource,
-        batchMemoryMappedIndex,
-        batchProcessingMode,
         storeEmptyColumns,
         encapsulatedTask,
         tmpStorageBytesPerTask
