+++ indexing-service/src/test/java/org/apache/druid/indexing/common/TaskToolboxTest.java
@@ -114,7 +114,6 @@ public class TaskToolboxTest
     TaskConfig taskConfig = new TaskConfigBuilder()
         .setBaseDir(temporaryFolder.newFile().toString())
         .setDefaultRowFlushBoundary(50000)
-        .setBatchProcessingMode(TaskConfig.BATCH_PROCESSING_MODE_DEFAULT.name())
         .build();
 
     taskToolbox = new TaskToolboxFactory(
