+++ processing/src/main/java/org/apache/druid/timeline/partition/BuildingHashBasedNumberedShardSpec.java
@@ -98,8 +98,8 @@ public class BuildingHashBasedNumberedShardSpec implements BuildingShardSpec<Has
   @Override
   public <T> PartitionChunk<T> createChunk(T obj)
   {
-    // This method can be called in AppenderatorImpl to create a sinkTimeline.
-    // The sinkTimeline doesn't seem in use in batch ingestion, let's set 'chunks' to 0 for now.
+    // This method can be called in StreamAppenderator to create a sinkTimeline.
+    // The sinkTimeline isn't used in batch ingestion, let's set 'chunks' to 0 for now.
     // HashBasedNumberedShardSpec is using NumberedPartitionChunk, so we use it here too.
     return new NumberedPartitionChunk<>(partitionId, 0, obj);
   }
