+++ server/src/main/java/org/apache/druid/segment/realtime/appenderator/UnifiedIndexerAppenderatorsManager.java
@@ -205,7 +205,7 @@ public class UnifiedIndexerAppenderatorsManager implements AppenderatorsManager
   }
 
   @Override
-  public Appenderator createOfflineAppenderatorForTask(
+  public Appenderator createBatchAppenderatorForTask(
       String taskId,
       DataSchema schema,
       AppenderatorConfig config,
@@ -226,89 +226,7 @@ public class UnifiedIndexerAppenderatorsManager implements AppenderatorsManager
           DatasourceBundle::new
       );
 
-      Appenderator appenderator = Appenderators.createOffline(
-          taskId,
-          schema,
-          rewriteAppenderatorConfigMemoryLimits(config),
-          metrics,
-          dataSegmentPusher,
-          objectMapper,
-          indexIO,
-          wrapIndexMerger(indexMerger),
-          rowIngestionMeters,
-          parseExceptionHandler,
-          useMaxMemoryEstimates,
-          centralizedDatasourceSchemaConfig
-      );
-      datasourceBundle.addAppenderator(taskId, appenderator);
-      return appenderator;
-    }
-  }
-
-  @Override
-  public Appenderator createOpenSegmentsOfflineAppenderatorForTask(
-      String taskId,
-      DataSchema schema,
-      AppenderatorConfig config,
-      SegmentGenerationMetrics metrics,
-      DataSegmentPusher dataSegmentPusher,
-      ObjectMapper objectMapper,
-      IndexIO indexIO,
-      IndexMerger indexMerger,
-      RowIngestionMeters rowIngestionMeters,
-      ParseExceptionHandler parseExceptionHandler,
-      boolean useMaxMemoryEstimates,
-      CentralizedDatasourceSchemaConfig centralizedDatasourceSchemaConfig
-  )
-  {
-    synchronized (this) {
-      DatasourceBundle datasourceBundle = datasourceBundles.computeIfAbsent(
-          schema.getDataSource(),
-          DatasourceBundle::new
-      );
-
-      Appenderator appenderator = Appenderators.createOpenSegmentsOffline(
-          taskId,
-          schema,
-          rewriteAppenderatorConfigMemoryLimits(config),
-          metrics,
-          dataSegmentPusher,
-          objectMapper,
-          indexIO,
-          wrapIndexMerger(indexMerger),
-          rowIngestionMeters,
-          parseExceptionHandler,
-          useMaxMemoryEstimates,
-          centralizedDatasourceSchemaConfig
-      );
-      datasourceBundle.addAppenderator(taskId, appenderator);
-      return appenderator;
-    }
-  }
-
-  @Override
-  public Appenderator createClosedSegmentsOfflineAppenderatorForTask(
-      String taskId,
-      DataSchema schema,
-      AppenderatorConfig config,
-      SegmentGenerationMetrics metrics,
-      DataSegmentPusher dataSegmentPusher,
-      ObjectMapper objectMapper,
-      IndexIO indexIO,
-      IndexMerger indexMerger,
-      RowIngestionMeters rowIngestionMeters,
-      ParseExceptionHandler parseExceptionHandler,
-      boolean useMaxMemoryEstimates,
-      CentralizedDatasourceSchemaConfig centralizedDatasourceSchemaConfig
-  )
-  {
-    synchronized (this) {
-      DatasourceBundle datasourceBundle = datasourceBundles.computeIfAbsent(
-          schema.getDataSource(),
-          DatasourceBundle::new
-      );
-
-      Appenderator appenderator = Appenderators.createClosedSegmentsOffline(
+      Appenderator appenderator = Appenderators.createBatch(
           taskId,
           schema,
           rewriteAppenderatorConfigMemoryLimits(config),
@@ -595,7 +513,7 @@ public class UnifiedIndexerAppenderatorsManager implements AppenderatorsManager
   /**
    * This wrapper around IndexMerger limits concurrent calls to the merge/persist methods used by
    * {@link StreamAppenderator} with a shared executor service. Merge/persist methods that are not used by
-   * AppenderatorImpl will throw an exception if called.
+   * StreamAppenderator will throw an exception if called.
    */
   public static class LimitedPoolIndexMerger implements IndexMerger
   {
