+++ server/src/test/java/org/apache/druid/segment/realtime/appenderator/BatchAppenderatorDriverTest.java
@@ -29,25 +29,30 @@ import org.apache.druid.indexing.overlord.SegmentPublishResult;
 import org.apache.druid.java.util.common.DateTimes;
 import org.apache.druid.java.util.common.Intervals;
 import org.apache.druid.java.util.common.granularity.Granularities;
+import org.apache.druid.java.util.common.granularity.Granularity;
 import org.apache.druid.segment.loading.DataSegmentKiller;
 import org.apache.druid.segment.realtime.appenderator.BaseAppenderatorDriver.SegmentsForSequence;
 import org.apache.druid.segment.realtime.appenderator.SegmentWithState.SegmentState;
-import org.apache.druid.segment.realtime.appenderator.StreamAppenderatorDriverTest.TestSegmentAllocator;
 import org.apache.druid.timeline.partition.NumberedShardSpec;
 import org.easymock.EasyMock;
 import org.easymock.EasyMockSupport;
+import org.joda.time.DateTime;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
 import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-public class OpenAndClosedSegmentsBatchAppenderatorDriverTest extends EasyMockSupport
+public class BatchAppenderatorDriverTest extends EasyMockSupport
 {
   private static final String DATA_SOURCE = "foo";
   private static final String VERSION = "abc123";
@@ -73,7 +78,7 @@ public class OpenAndClosedSegmentsBatchAppenderatorDriverTest extends EasyMockSu
   );
 
   private SegmentAllocator allocator;
-  private OpenAndClosedSegmentsAppenderatorTester openAndClosedSegmentsAppenderatorTester;
+  private BatchAppenderatorTester appenderatorTester;
   private BatchAppenderatorDriver driver;
   private DataSegmentKiller dataSegmentKiller;
 
@@ -84,16 +89,13 @@ public class OpenAndClosedSegmentsBatchAppenderatorDriverTest extends EasyMockSu
   @Before
   public void setup()
   {
-    openAndClosedSegmentsAppenderatorTester =
-        new OpenAndClosedSegmentsAppenderatorTester(MAX_ROWS_IN_MEMORY, false,
-                                                    false
-        );
+    appenderatorTester = new BatchAppenderatorTester(MAX_ROWS_IN_MEMORY);
     allocator = new TestSegmentAllocator(DATA_SOURCE, Granularities.HOUR);
     dataSegmentKiller = createStrictMock(DataSegmentKiller.class);
     driver = new BatchAppenderatorDriver(
-        openAndClosedSegmentsAppenderatorTester.getAppenderator(),
+        appenderatorTester.getAppenderator(),
         allocator,
-        new TestPublishedSegmentRetriever(openAndClosedSegmentsAppenderatorTester.getPushedSegments()),
+        new TestPublishedSegmentRetriever(appenderatorTester.getPushedSegments()),
         dataSegmentKiller
     );
 
@@ -109,7 +111,7 @@ public class OpenAndClosedSegmentsBatchAppenderatorDriverTest extends EasyMockSu
     driver.close();
   }
 
-  @Test (timeout = 2000L)
+  @Test
   public void testSimple() throws Exception
   {
     Assert.assertNull(driver.startJob(null));
@@ -125,7 +127,8 @@ public class OpenAndClosedSegmentsBatchAppenderatorDriverTest extends EasyMockSu
     checkSegmentStates(2, SegmentState.PUSHED_AND_DROPPED);
 
     final SegmentsAndCommitMetadata published =
-        driver.publishAll(null, null, makeOkPublisher(), Function.identity(), null).get(TIMEOUT, TimeUnit.MILLISECONDS);
+        driver.publishAll(null, Collections.emptySet(), makeOkPublisher(), Function.identity(), null)
+              .get(TIMEOUT, TimeUnit.MILLISECONDS);
 
     Assert.assertEquals(
         ImmutableSet.of(
@@ -159,7 +162,8 @@ public class OpenAndClosedSegmentsBatchAppenderatorDriverTest extends EasyMockSu
     }
 
     final SegmentsAndCommitMetadata published =
-        driver.publishAll(null, null, makeOkPublisher(), Function.identity(), null).get(TIMEOUT, TimeUnit.MILLISECONDS);
+        driver.publishAll(null, Collections.emptySet(), makeOkPublisher(), Function.identity(), null)
+              .get(TIMEOUT, TimeUnit.MILLISECONDS);
 
     Assert.assertEquals(
         ImmutableSet.of(
@@ -181,7 +185,7 @@ public class OpenAndClosedSegmentsBatchAppenderatorDriverTest extends EasyMockSu
   {
     Assert.assertNull(driver.startJob(null));
     driver.close();
-    openAndClosedSegmentsAppenderatorTester.getAppenderator().close();
+    appenderatorTester.getAppenderator().close();
 
     Assert.assertNull(driver.startJob(null));
   }
@@ -202,4 +206,38 @@ public class OpenAndClosedSegmentsBatchAppenderatorDriverTest extends EasyMockSu
   {
     return (segmentsToBeOverwritten, segmentsToPublish, commitMetadata, schema) -> SegmentPublishResult.ok(ImmutableSet.of());
   }
+
+  static class TestSegmentAllocator implements SegmentAllocator
+  {
+    private final String dataSource;
+    private final Granularity granularity;
+    private final Map<Long, AtomicInteger> counters = new HashMap<>();
+
+    public TestSegmentAllocator(String dataSource, Granularity granularity)
+    {
+      this.dataSource = dataSource;
+      this.granularity = granularity;
+    }
+
+    @Override
+    public SegmentIdWithShardSpec allocate(
+        final InputRow row,
+        final String sequenceName,
+        final String previousSegmentId,
+        final boolean skipSegmentLineageCheck
+    )
+    {
+      DateTime dateTimeTruncated = granularity.bucketStart(row.getTimestamp());
+      final long timestampTruncated = dateTimeTruncated.getMillis();
+      counters.putIfAbsent(timestampTruncated, new AtomicInteger());
+      final int partitionNum = counters.get(timestampTruncated).getAndIncrement();
+      return new SegmentIdWithShardSpec(
+          dataSource,
+          granularity.bucket(dateTimeTruncated),
+          VERSION,
+          new NumberedShardSpec(partitionNum, 0)
+      );
+    }
+  }
+
 }
